pyproject.toml
[tool.poetry]
name = "dex-trading"
version = "0.1.0"
description = ""
authors = ["Rick Albaugh <albaughrick@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asyncio = "^3.4.3"
DateTime = "^5.5"
solders = "^0.21.0"
construct = "2.10.68"
solana = "^0.33.0"
pandas = "^2.2.1"
tabulate = "^0.9.0"
asyncstdlib = "^3.12.2"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
base58 = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
